working: full find_min_cut rewrite
